refactor: Move out CredentialLoad
